страница rubric виджеты для страницы rubric npm install --save @ckeditor/ckeditor5-build-classic

+++ database/migrations/2024_07_24_045151_create_credits_logs_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('credits_logs', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->integer('credits')->nullable();
+            $table->smallInteger('credits')->nullable();
             $table->unsignedBigInteger('bb_id')->nullable();
             $table->string('description')->nullable();
             $table->foreign('bb_id')->references('id')->on('bbs')->cascadeOnDelete();
